Move CssModuleComposesIssue validation into CssModuleComposeReference::resolve_reference
Instead of validating in module_references(), move the logic into the
reference itself. Introduce a CssModuleComposable marker trait in
compose.rs; EcmascriptCssModule implements it. resolve_reference() now
validates both cases (unresolvable + non-CSS-module target) and emits
CssModuleComposesIssue directly, making the issue available during
module graph traversal (e.g. get_compilation_issues) without coupling
the validation to the consumer (module_references or codegen).
